Refactor: alias imports, rename atom module, discriminated union for ReceiptAnalysisResult
- Replace relative imports with @/ aliases in imagePrepStateAtoms, EraseRectangle, EraseTool
- Rename pendingImageAtom.ts → imagePrepAtoms.ts (both atoms in one neutral-named module)
- Update all import sites (HomePage, ImagePrepPage, PreviewPage)
- Change ReceiptUploader import in HomePage from barrel to direct component path
- Replace ReceiptAnalysisResult broad interface with a discriminated union on success/is_receipt
- Remove unsafe `as string | undefined` cast in parseError; simplify receipt_data?.id guard
