+++ frontend/src/features/image-prep/atoms/imagePrepStateAtoms.ts
@@ -1,7 +1,7 @@
 import { atom } from 'jotai';
 import type { Crop, PixelCrop } from 'react-image-crop';
 
-import type { EraseRect, ImageDimensions } from '../utils/canvasOperations';
+import type { EraseRect, ImageDimensions } from '@/features/image-prep/utils/canvasOperations';
 
 export const cropAtom = atom<Crop | undefined>(undefined);
 export const completedCropAtom = atom<PixelCrop | null>(null);
